refactor(desktop): D38 split ChatFeed into hook, bubble, and dock modules
Extract useChatFeedController (state/effects/HITL/scroll/send), ChatMessageBubble
(hybrid disclosure, media parse, message rows), and ChatFeedDock (fixed input
stack). ChatFeed.tsx remains layout shell; configuredEngineToRuntimeLabel
re-export preserved for App.tsx.
